Refactor[Info]: moved getStatus function to a separate file
